it successfully puts the old_file into old_list and passes both old_list and old_num_clients to email template - although the old_list in the template is a list with each family name as a list -very minor

+++ reporter.py
@@ -1,6 +1,9 @@
+import os
+import os.path
 from selenium import webdriver
 
 from functions import (
+	old_file_csv_to_list,
 	kaiser_login,
 	initialize_list,
 	scroll_add_new_to_list,
@@ -8,11 +11,11 @@ from functions import (
 	send_the_email,
 	)
 
-os.rename("new_file.csv", "old_file.csv")
+if os.path.isfile("new_file.csv"):
+	os.rename("new_file.csv", "old_file.csv")
 
-# bring in old list as old_list
-
-# set len(old_list) to old_num_clients
+old_list = old_file_csv_to_list()
+old_num_clients = len(old_list)
 
 url = "https://clear.kaiserpermanente.org/?kp_shortcut_referrer=kp.org/clear#/login"
 driver = webdriver.Firefox()
@@ -35,13 +38,12 @@ new_num_clients = len(new_list)
 create_newFile(new_list)
 
 # number difference b/w old_num_clients & new_num_clients as num_client_difference
+
 # make list of added or removed clients
 
 # use variation of compare.py to show added and removed clients
 
-
-
 # LAST FEATURE: count the total number of clients by accessing the number of/
 # family members 
 
-send_the_email(new_list, new_num_clients)
+send_the_email(old_list, old_num_clients, new_list, new_num_clients)
